Add expiry date to standard document
This commit restructure some code, and it also do some changes in
the way to find a document. Now, it will only use the attributes
user has supplied to find a standard. This also normalize some of
the attributes, so we don't store duplicates copies for multiple
strings.

This commit add also adds `expires_at` date to the `Standard`
table. When storing a new document, then by default it will set
15 days validity, but if those are the specific use cases then
it will only set it for 5 days.

Related: #10
