+++ db/migrate/20191010102606_add_expires_at_to_standards.rb
@@ -1,0 +1,5 @@
+class AddExpiresAtToStandards < ActiveRecord::Migration[5.2]
+  def change
+    add_column :standards, :expires_at, :datetime
+  end
+end
