+++ spec/models/standard_finder_spec.rb
@@ -1,35 +1,36 @@
 require "rails_helper"
 
 RSpec.describe StandardFinder do
-  describe ".find_by" do
-    context "with valid standard details" do
+  describe ".find_or_fetch_by" do
+    context "with existing standard" do
       it "finds and returns the standard" do
-        year = 2003
-        code = "ISO 19115"
-        stub_relaton_document(code, year)
+        standard = create(:standard, :with_document)
 
-        standard = StandardFinder.find_by(code: code, year: year)
-        hash_document = standard.document_in_hash
+        found_standard = StandardFinder.find_or_fetch_by(
+          code: standard.code, year: standard.year,
+        )
 
-        expect(standard.code).to eq(code)
-        expect(standard.year).to eq(year)
-        expect(standard.standard_type).to eq("iso")
-        expect(standard.document_number).to eq("19115")
-        expect(hash_document["date"]["value"]).to eq("2003-01-01")
-        expect(hash_document["docid"]["id"]).to eq("ISO 19115:2003")
-        expect(standard.document_in_xml).to include('<bibitem id="ISO19115-200')
+        expect(found_standard).to eq(standard)
+        expect(found_standard.year).to eq(standard.year)
       end
     end
 
-    context "with invalid standard details" do
-      it "does not returns anything but nil" do
-        year = 2019
-        code = "ISO 19115"
+    context "with non existing document" do
+      it "fetch and returns the new document" do
+        standard_fixture = build(:standard)
+        stub_relaton_document(standard_fixture.code, standard_fixture.year)
 
-        stub_relaton_invalid_document(code, year)
-        standard = StandardFinder.find_by(code: code, year: year)
+        standard = StandardFinder.find_or_fetch_by(
+          code: standard_fixture.code, year: standard_fixture.year,
+        )
 
-        expect(standard).to be_nil
+        expect(standard.code).to eq(standard_fixture.code)
+        expect(standard.year).to eq(standard_fixture.year)
+        expect(standard.expires_at.to_date).to eq(5.days.from_now.to_date)
+        expect(standard.standard_type).to eq(standard_fixture.standard_type)
+        expect(standard.document_in_json["date"]["value"]).to eq("2003-01-01")
+        expect(standard.document_in_xml).to include('<bibitem id="ISO19115-200')
+        expect(standard.document_number).to eq(standard_fixture.document_number)
       end
     end
   end
